requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
ngrok==0.0.1
ordereddict==1.1
PyJWT==1.7.1
pylint==2.3.1
PySocks==1.7.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
six==1.12.0
twilio==6.29.1
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
